# api/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Check-In API",
        "contact": {},
        "license": {
            "name": "GPL-3.0"
        },
        "version": "1.0"
    },
    "paths": {
        "/auth/refresh": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Refresh access token",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorDto"
                        }
                    }
                }
            }
        },
        "/auth/signin": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in a user",
                "parameters": [
                    {
                        "description": "SignInDto",
                        "name": "signInDto",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SignInDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ErrorDto"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorDto"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ErrorDto"
                        }
                    }
                }
            }
        },
        "/auth/signout": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign out a user",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/ErrorDto"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ErrorDto": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {},
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "SignInDto": {
            "type": "object",
            "properties": {
                "password": {
                    "type": "string"
                },
                "rememberMe": {
                    "type": "boolean"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    }
}